+++ src/helpers/createMetadataEventDetailPage.ts
@@ -2,10 +2,10 @@ import urlcat from 'urlcat';
 
 import { SITE_URL } from '@/constants/index.js';
 import { createSiteMetadata } from '@/helpers/createSiteMetadata.js';
-import { getFireflyActivityInfo } from '@/services/getFireflyActivityInfo.js';
+import { FireflyActivityProvider } from '@/providers/firefly/Activity.js';
 
 export async function createMetadataEventDetailPage(eventName: string) {
-    const info = await getFireflyActivityInfo(eventName);
+    const info = await FireflyActivityProvider.getFireflyActivityInfo(eventName);
     const title = info.title;
     const description = info.sub_title;
     const images = [info.open_graph_url];
